requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
boto3==1.35.64
botocore==1.35.64
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
cssselect2==0.7.0
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
html5lib==1.1
idna==3.9
jmespath==1.0.1
lxml==5.3.0
numpy==1.26.3
oscrypto==1.3.0
pandas==2.1.4
pillow==10.4.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
s3transfer==0.10.3
setuptools==75.0.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.16
django-filters
psycopg2-binary
